Do not attempt to call 'gradle testClasses'
This is a legacy from before the Gradle plugin. It was called to
compile the test code. However, 'gradle cifuzzPrintTestClasspath',
which we call next, will take care of compiling the code. This
will select the correct compile task(s) depending on which test
source set was selected etc.

'testClasses' might also not exist in all projects. Especially Android
projects usually do not have it.